feat(): fetchUtils.ts
